tasks: split common omnibus code in its own file
