+++ docs/dev/agent_omnibus.md
@@ -16,10 +16,10 @@ Use the provided Docker images to build a DEB or RPM
 package for Linux. You need to have Docker already running on your machine.
 
 From the `datadog-agent` source folder, use the following command to run the
-`agent.omnibus-build` task in a Docker container:
+`omnibus.build` task in a Docker container:
 
 ```
-docker run -v "$PWD:/go/src/github.com/DataDog/datadog-agent" -v "/tmp/omnibus:/omnibus" -v "/tmp/opt/datadog-agent:/opt/datadog-agent" -v"/tmp/gems:/gems" --workdir=/go/src/github.com/DataDog/datadog-agent datadog/agent-buildimages-deb_x64 inv -e agent.omnibus-build --base-dir=/omnibus --gem-path=/gems
+docker run -v "$PWD:/go/src/github.com/DataDog/datadog-agent" -v "/tmp/omnibus:/omnibus" -v "/tmp/opt/datadog-agent:/opt/datadog-agent" -v"/tmp/gems:/gems" --workdir=/go/src/github.com/DataDog/datadog-agent datadog/agent-buildimages-deb_x64 inv -e omnibus.build --base-dir=/omnibus --gem-path=/gems
 ```
 
 The container will share 3 volumes with the host to avoid starting from scratch
@@ -60,20 +60,20 @@ the filesystem without disrupting anything.
 To run Omnibus and build the package, make the `/opt` folder world readable and run:
 
 ```
-inv agent.omnibus-build --base-dir=$HOME/.omnibus
+inv omnibus.build --base-dir=$HOME/.omnibus
 ```
 
 On Mac, you might want to skip the signing step by running:
 
 ```
-inv agent.omnibus-build --base-dir=$HOME/.omnibus --skip-sign
+inv omnibus.build --base-dir=$HOME/.omnibus --skip-sign
 ```
 
 The path you pass with the `--base-dir` option will contain the sources
 downloaded by Omnibus in the `src` folder, the binaries cached after building
 those sources in the `cache` folder and the final deb/rpm/dmg artifacts in the
 `pkg` folder. You can fine tune an Omnibus run passing more options, see
-`inv agent.omnibus-build --help` for the list of all the available options.
+`inv omnibus.build --help` for the list of all the available options.
 
 **Note:** it's strongly advised to pass `--base-dir` and point to a directory
 outside the Agent repo. By default Omnibus stores packages in the project folder
@@ -87,8 +87,8 @@ To build on Windows, [Docker Desktop](https://docs.docker.com/docker-for-windows
 
 Start a Powershell prompt and navigate to your local clone of the `datadog-agent` repo.
 
- Run the following command: 
- 
+ Run the following command:
+
 ```powershell
 docker run -v "$(Get-Location):c:\mnt" -e OMNIBUS_TARGET=main -e RELEASE_VERSION=nightly -e MAJOR_VERSION=7 -e PY_RUNTIMES=3 -e TARGET_ARCH=x64 datadog/agent-buildimages-windows_x64:1809 c:\mnt\tasks\winbuildscripts\buildwin.bat
 ```
