+++ tasks/agent.py
@@ -4,16 +4,13 @@ Agent namespaced tasks
 
 import ast
 import glob
-import json
 import os
 import platform
 import re
 import shutil
 import sys
 import tempfile
-from datetime import datetime
 
-import requests
 from invoke import task
 from invoke.exceptions import Exit, ParseError
 
@@ -29,13 +26,10 @@ from tasks.libs.common.utils import (
     get_goenv,
     get_version,
     has_both_python,
-    load_release_versions,
-    timed,
 )
 from tasks.rtloader import clean as rtloader_clean
 from tasks.rtloader import install as rtloader_install
 from tasks.rtloader import make as rtloader_make
-from tasks.ssm import get_pfx_pass, get_signing_cert
 from tasks.windows_resources import build_messagetable, build_rc, versioninfo_vars
 
 # constants
@@ -413,7 +407,7 @@ def image_build(ctx, arch='amd64', base_dir="omnibus", python_version="2", skip_
     # get the last debian package built
     if not list_of_files:
         print(f"No debian package build found in {pkg_dir}")
-        print("See agent.omnibus-build")
+        print("See omnibus.build")
         raise Exit(code=1)
     latest_file = max(list_of_files, key=os.path.getctime)
     shutil.copy2(latest_file, build_context)
@@ -620,401 +614,6 @@ def _linux_integration_tests(ctx, race=False, remote_docker=False, go_mod="mod",
 
     for prefix in prefixes:
         ctx.run(f"{go_cmd} {prefix}")
-
-
-def get_omnibus_env(
-    ctx,
-    skip_sign=False,
-    release_version="nightly",
-    major_version='7',
-    python_runtimes='3',
-    hardened_runtime=False,
-    system_probe_bin=None,
-    go_mod_cache=None,
-    flavor=AgentFlavor.base,
-    pip_config_file="pip.conf",
-):
-    env = load_release_versions(ctx, release_version)
-
-    # If the host has a GOMODCACHE set, try to reuse it
-    if not go_mod_cache and os.environ.get('GOMODCACHE'):
-        go_mod_cache = os.environ.get('GOMODCACHE')
-
-    if go_mod_cache:
-        env['OMNIBUS_GOMODCACHE'] = go_mod_cache
-
-    if int(major_version) > 6:
-        env['OMNIBUS_OPENSSL_SOFTWARE'] = 'openssl3'
-
-    env_override = ['INTEGRATIONS_CORE_VERSION', 'OMNIBUS_SOFTWARE_VERSION']
-    for key in env_override:
-        value = os.environ.get(key)
-        # Only overrides the env var if the value is a non-empty string.
-        if value:
-            env[key] = value
-
-    if sys.platform == 'win32' and os.environ.get('SIGN_WINDOWS'):
-        # get certificate and password from ssm
-        pfxfile = get_signing_cert(ctx)
-        pfxpass = get_pfx_pass(ctx)
-        env['SIGN_PFX'] = str(pfxfile)
-        env['SIGN_PFX_PW'] = str(pfxpass)
-
-    if sys.platform == 'darwin':
-        # Target MacOS 10.12
-        env['MACOSX_DEPLOYMENT_TARGET'] = '10.12'
-
-    if skip_sign:
-        env['SKIP_SIGN_MAC'] = 'true'
-    if hardened_runtime:
-        env['HARDENED_RUNTIME_MAC'] = 'true'
-
-    env['PACKAGE_VERSION'] = get_version(
-        ctx, include_git=True, url_safe=True, major_version=major_version, include_pipeline_id=True
-    )
-    env['MAJOR_VERSION'] = major_version
-    env['PY_RUNTIMES'] = python_runtimes
-
-    # Since omnibus and the invoke task won't run in the same folder
-    # we need to input the absolute path of the pip config file
-    env['PIP_CONFIG_FILE'] = os.path.abspath(pip_config_file)
-
-    if system_probe_bin:
-        env['SYSTEM_PROBE_BIN'] = system_probe_bin
-    env['AGENT_FLAVOR'] = flavor.name
-
-    # We need to override the workers variable in omnibus build when running on Kubernetes runners,
-    # otherwise, ohai detect the number of CPU on the host and run the make jobs with all the CPU.
-    if os.environ.get('KUBERNETES_CPU_REQUEST'):
-        env['OMNIBUS_WORKERS_OVERRIDE'] = str(int(os.environ.get('KUBERNETES_CPU_REQUEST')) + 1)
-    # Forward the DEPLOY_AGENT variable so that we can use a higher compression level for deployed artifacts
-    if os.environ.get('DEPLOY_AGENT'):
-        env['DEPLOY_AGENT'] = os.environ.get('DEPLOY_AGENT')
-    if 'PACKAGE_ARCH' in os.environ:
-        env['PACKAGE_ARCH'] = os.environ.get('PACKAGE_ARCH')
-
-    return env
-
-
-def omnibus_run_task(
-    ctx, task, target_project, base_dir, env, omnibus_s3_cache=False, log_level="info", host_distribution=None
-):
-    with ctx.cd("omnibus"):
-        overrides_cmd = ""
-        if base_dir:
-            overrides_cmd = f"--override=base_dir:{base_dir}"
-        if host_distribution:
-            overrides_cmd += f" --override=host_distribution:{host_distribution}"
-
-        omnibus = "bundle exec omnibus"
-        if sys.platform == 'win32':
-            omnibus = "bundle exec omnibus.bat"
-        elif sys.platform == 'darwin':
-            # HACK: This is an ugly hack to fix another hack made by python3 on MacOS
-            # The full explanation is available on this PR: https://github.com/DataDog/datadog-agent/pull/5010.
-            omnibus = "unset __PYVENV_LAUNCHER__ && bundle exec omnibus"
-
-        if omnibus_s3_cache:
-            populate_s3_cache = "--populate-s3-cache"
-        else:
-            populate_s3_cache = ""
-
-        cmd = "{omnibus} {task} {project_name} --log-level={log_level} {populate_s3_cache} {overrides}"
-        args = {
-            "omnibus": omnibus,
-            "task": task,
-            "project_name": target_project,
-            "log_level": log_level,
-            "overrides": overrides_cmd,
-            "populate_s3_cache": populate_s3_cache,
-        }
-
-        ctx.run(cmd.format(**args), env=env)
-
-
-def bundle_install_omnibus(ctx, gem_path=None, env=None, max_try=2):
-    with ctx.cd("omnibus"):
-        # make sure bundle install starts from a clean state
-        try:
-            os.remove("Gemfile.lock")
-        except Exception:
-            pass
-
-        cmd = "bundle install"
-        if gem_path:
-            cmd += f" --path {gem_path}"
-
-        for trial in range(max_try):
-            res = ctx.run(cmd, env=env, warn=True)
-            if res.ok:
-                return
-            if not should_retry_bundle_install(res):
-                return
-            print(f"Retrying bundle install, attempt {trial + 1}/{max_try}")
-
-
-def should_retry_bundle_install(res):
-    # We sometimes get a Net::HTTPNotFound error when fetching the
-    # license-scout gem. This is a transient error, so we retry the bundle install
-    if "Net::HTTPNotFound:" in res.stderr:
-        return True
-    return False
-
-
-def _send_build_metrics(ctx, overall_duration):
-    # We only want to generate those metrics from the CI
-    if sys.platform == 'win32':
-        src_dir = "C:/buildroot/datadog-agent"
-        aws_cmd = "aws.cmd"
-    else:
-        src_dir = os.environ.get('CI_PROJECT_DIR')
-        aws_cmd = "aws"
-    job_name = os.environ.get('CI_JOB_NAME_SLUG')
-    branch = os.environ.get('CI_COMMIT_REF_NAME')
-    pipeline_id = os.environ.get('CI_PIPELINE_ID')
-    if not job_name or not branch or not src_dir or not pipeline_id:
-        print(
-            '''Missing required environment variables, this is probably not a CI job.
-                  skipping sending build metrics'''
-        )
-        return
-
-    series = []
-    timestamp = int(datetime.now().timestamp())
-    with open(f'{src_dir}/omnibus/pkg/build-summary.json') as summary_json:
-        j = json.load(summary_json)
-        # Various software build durations are all sent as the `datadog.agent.build.duration` metric
-        # with a specific tag for each software.
-        for software, metrics in j['build'].items():
-            series.append(
-                {
-                    'metric': 'datadog.agent.build.duration',
-                    'points': [{'timestamp': timestamp, 'value': metrics['build_duration']}],
-                    'tags': [
-                        f'software:{software}',
-                        f'cached:{metrics["cached"]}',
-                        f'job:{job_name}',
-                        f'branch:{branch}',
-                        f'pipeline:{pipeline_id}',
-                    ],
-                    'unit': 'seconds',
-                    'type': 0,
-                }
-            )
-        # We also provide the total duration for the omnibus build as a separate metric
-        series.append(
-            {
-                'metric': 'datadog.agent.build.total',
-                'points': [{'timestamp': timestamp, 'value': overall_duration}],
-                'tags': [
-                    f'job:{job_name}',
-                    f'branch:{branch}',
-                    f'pipeline:{pipeline_id}',
-                ],
-                'unit': 'seconds',
-                'type': 0,
-            }
-        )
-        # Stripping might not always be enabled so we conditionally read the metric
-        if "strip" in j:
-            series.append(
-                {
-                    'metric': 'datadog.agent.build.strip',
-                    'points': [{'timestamp': timestamp, 'value': j['strip']}],
-                    'tags': [
-                        f'job:{job_name}',
-                        f'branch:{branch}',
-                        f'pipeline:{pipeline_id}',
-                    ],
-                    'unit': 'seconds',
-                    'type': 0,
-                }
-            )
-        # And all packagers duration as another separated metric
-        for packager, duration in j['packaging'].items():
-            series.append(
-                {
-                    'metric': 'datadog.agent.package.duration',
-                    'points': [{'timestamp': timestamp, 'value': duration}],
-                    'tags': [
-                        f'job:{job_name}',
-                        f'branch:{branch}',
-                        f'packager:{packager}',
-                        f'pipeline:{pipeline_id}',
-                    ],
-                    'unit': 'seconds',
-                    'type': 0,
-                }
-            )
-    dd_api_key = ctx.run(
-        f'{aws_cmd} ssm get-parameter --region us-east-1 --name {os.environ["API_KEY_ORG2_SSM_NAME"]} --with-decryption --query "Parameter.Value" --out text',
-        hide=True,
-    ).stdout.strip()
-    headers = {'Accept': 'application/json', 'Content-Type': 'application/json', 'DD-API-KEY': dd_api_key}
-    r = requests.post("https://api.datadoghq.com/api/v2/series", json={'series': series}, headers=headers)
-    if r.ok:
-        print('Successfully sent build metrics to DataDog')
-    else:
-        print(f'Failed to send build metrics to DataDog: {r.status_code}')
-        print(r.text)
-
-
-# hardened-runtime needs to be set to False to build on MacOS < 10.13.6, as the -o runtime option is not supported.
-@task(
-    help={
-        'skip-sign': "On macOS, use this option to build an unsigned package if you don't have Datadog's developer keys.",
-        'hardened-runtime': "On macOS, use this option to enforce the hardened runtime setting, adding '-o runtime' to all codesign commands",
-    }
-)
-def omnibus_build(
-    ctx,
-    flavor=AgentFlavor.base.name,
-    agent_binaries=False,
-    log_level="info",
-    base_dir=None,
-    gem_path=None,
-    skip_deps=False,
-    skip_sign=False,
-    release_version="nightly",
-    major_version='7',
-    python_runtimes='3',
-    omnibus_s3_cache=False,
-    hardened_runtime=False,
-    system_probe_bin=None,
-    go_mod_cache=None,
-    python_mirror=None,
-    pip_config_file="pip.conf",
-    host_distribution=None,
-):
-    """
-    Build the Agent packages with Omnibus Installer.
-    """
-
-    flavor = AgentFlavor[flavor]
-    if not skip_deps:
-        with timed(quiet=True) as deps_elapsed:
-            deps(ctx)
-
-    # base dir (can be overridden through env vars, command line takes precedence)
-    base_dir = base_dir or os.environ.get("OMNIBUS_BASE_DIR")
-
-    if base_dir is not None and sys.platform == 'win32':
-        # On Windows, prevent backslashes in the base_dir path otherwise omnibus will fail with
-        # error 'no matched files for glob copy' at the end of the build.
-        base_dir = base_dir.replace(os.path.sep, '/')
-
-    env = get_omnibus_env(
-        ctx,
-        skip_sign=skip_sign,
-        release_version=release_version,
-        major_version=major_version,
-        python_runtimes=python_runtimes,
-        hardened_runtime=hardened_runtime,
-        system_probe_bin=system_probe_bin,
-        go_mod_cache=go_mod_cache,
-        flavor=flavor,
-        pip_config_file=pip_config_file,
-    )
-
-    target_project = "agent"
-    if flavor.is_iot():
-        target_project = "iot-agent"
-    elif agent_binaries:
-        target_project = "agent-binaries"
-
-    # Get the python_mirror from the PIP_INDEX_URL environment variable if it is not passed in the args
-    python_mirror = python_mirror or os.environ.get("PIP_INDEX_URL")
-
-    # If a python_mirror is set then use it for pip by adding it in the pip.conf file
-    pip_index_url = f"[global]\nindex-url = {python_mirror}" if python_mirror else ""
-
-    # We're passing the --index-url arg through a pip.conf file so that omnibus doesn't leak the token
-    with open(pip_config_file, 'w') as f:
-        f.write(pip_index_url)
-
-    with timed(quiet=True) as bundle_elapsed:
-        bundle_install_omnibus(ctx, gem_path, env)
-
-    with timed(quiet=True) as omnibus_elapsed:
-        omnibus_run_task(
-            ctx=ctx,
-            task="build",
-            target_project=target_project,
-            base_dir=base_dir,
-            env=env,
-            omnibus_s3_cache=omnibus_s3_cache,
-            log_level=log_level,
-            host_distribution=host_distribution,
-        )
-
-    # Delete the temporary pip.conf file once the build is done
-    os.remove(pip_config_file)
-
-    print("Build component timing:")
-    if not skip_deps:
-        print(f"Deps:    {deps_elapsed.duration}")
-    print(f"Bundle:  {bundle_elapsed.duration}")
-    print(f"Omnibus: {omnibus_elapsed.duration}")
-    _send_build_metrics(ctx, omnibus_elapsed.duration)
-
-
-@task
-def omnibus_manifest(
-    ctx,
-    platform=None,
-    arch=None,
-    flavor=AgentFlavor.base.name,
-    agent_binaries=False,
-    log_level="info",
-    base_dir=None,
-    gem_path=None,
-    skip_sign=False,
-    release_version="nightly",
-    major_version='7',
-    python_runtimes='3',
-    hardened_runtime=False,
-    system_probe_bin=None,
-    go_mod_cache=None,
-):
-    flavor = AgentFlavor[flavor]
-    # base dir (can be overridden through env vars, command line takes precedence)
-    base_dir = base_dir or os.environ.get("OMNIBUS_BASE_DIR")
-
-    env = get_omnibus_env(
-        ctx,
-        skip_sign=skip_sign,
-        release_version=release_version,
-        major_version=major_version,
-        python_runtimes=python_runtimes,
-        hardened_runtime=hardened_runtime,
-        system_probe_bin=system_probe_bin,
-        go_mod_cache=go_mod_cache,
-        flavor=flavor,
-    )
-
-    target_project = "agent"
-    if flavor.is_iot():
-        target_project = "iot-agent"
-    elif agent_binaries:
-        target_project = "agent-binaries"
-
-    bundle_install_omnibus(ctx, gem_path, env)
-
-    task = "manifest"
-    if platform is not None:
-        task += f" --platform-family={platform} --platform={platform} "
-    if arch is not None:
-        task += f" --architecture={arch} "
-
-    omnibus_run_task(
-        ctx=ctx,
-        task=task,
-        target_project=target_project,
-        base_dir=base_dir,
-        env=env,
-        omnibus_s3_cache=False,
-        log_level=log_level,
-    )
 
 
 @task
