+++ tasks/winbuildscripts/dobuild.bat
@@ -9,11 +9,11 @@ if NOT DEFINED MAJOR_VERSION set MAJOR_VERSION=%~2
 if NOT DEFINED PY_RUNTIMES set PY_RUNTIMES=%~3
 if NOT DEFINED GO_VERSION_CHECK set GO_VERSION_CHECK=%~4
 
-set OMNIBUS_BUILD=agent.omnibus-build
+set OMNIBUS_BUILD=omnibus.build
 set OMNIBUS_ARGS=--python-runtimes "%PY_RUNTIMES%"
 
 if "%OMNIBUS_TARGET%" == "iot" set OMNIBUS_ARGS=--flavor iot
-if "%OMNIBUS_TARGET%" == "dogstatsd" set OMNIBUS_BUILD=dogstatsd.omnibus-build && set OMNIBUS_ARGS=
+if "%OMNIBUS_TARGET%" == "dogstatsd" set OMNIBUS_ARGS=--target-project dogstatsd
 if "%OMNIBUS_TARGET%" == "agent_binaries" set OMNIBUS_ARGS=%OMNIBUS_ARGS% --agent-binaries
 if DEFINED GOMODCACHE set OMNIBUS_ARGS=%OMNIBUS_ARGS% --go-mod-cache %GOMODCACHE%
 if DEFINED USE_S3_CACHING set OMNIBUS_ARGS=%OMNIBUS_ARGS% %USE_S3_CACHING%
